tools/substack/requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0.0
# install the submodule in editable mode:
#   pip install -e ../../external/python-substack
